CHANGE page dir to app dir for collection, new collection, collection details, components folders
