+++ app/collection/[id]/not-found.tsx
@@ -1,0 +1,5 @@
+export default function CollectionNotFound() {
+  return (
+    <div>Collection doesn't exist</div>
+  )
+}
