+++ app/layout.tsx
@@ -1,0 +1,34 @@
+import Nav from "../components/layouts/Nav";
+
+import "../styles/globals.css";
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <head>
+        <meta charSet="UTF-8" />
+        <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" />
+        <link
+          href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700;800&display=swap"
+          rel="stylesheet"
+        />
+        <link rel="icon" type="image/x-icon" href="/js.png" />
+        <title>Jmember</title>
+      </head>
+
+      <body>
+        <div className="min-h-screen bg-gray-900 text-gray-200">
+          <Nav />
+          {children}
+        </div>
+      </body>
+    </html>
+  );
+}
