+++ app/loading.tsx
@@ -1,0 +1,5 @@
+export default function loading() {
+  return (
+    <div>loading</div>
+  )
+}
