+++ components/lesson/EditButton.tsx
@@ -1,7 +1,9 @@
+import Link from "next/link";
+
 export default function EditButton({ id }: { id: string }) {
   return (
-    <a
-      href={`${id}/edit`}
+    <Link 
+      href={`/collection/${id}/edit`}
       className="mr-2 mb-2 block rounded-lg bg-gray-600 px-5 py-2.5 text-sm font-medium text-white hover:bg-gray-700 focus:outline-none"
     >
       <svg
@@ -18,6 +20,6 @@ export default function EditButton({ id }: { id: string }) {
           d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L6.832 19.82a4.5 4.5 0 01-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 011.13-1.897L16.863 4.487zm0 0L19.5 7.125"
         />
       </svg>
-    </a>
+    </Link>
   );
 }
