+++ components/lesson/Flashcard.tsx
