+++ components/lesson/NextCard.tsx
