+++ package.json
@@ -10,24 +10,25 @@
     "preinstall": "npx only-allow pnpm"
   },
   "dependencies": {
+    "@next/font": "^13.0.2",
     "@react-query-firebase/firestore": "^1.0.0-dev.7",
-    "@types/node": "18.11.9",
-    "@types/react": "18.0.25",
-    "@types/react-dom": "18.0.8",
     "dotenv": "^16.0.3",
-    "dotenv-webpack": "^8.0.1",
     "firebase": "^9.13.0",
     "next": "13.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-firebase-hooks": "^5.0.3",
-    "typescript": "4.8.4"
+    "react-query": "^3.23.2"
   },
   "devDependencies": {
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.18",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
-    "tailwindcss": "^3.2.2"
+    "tailwindcss": "^3.2.2",
+    "typescript": "^4.8.4"
   }
 }
