--- pages/components/HeadTag.tsx
@@ -1,17 +1,0 @@
-import Head from "next/head";
-
-export default function HeadTag() {
-  return (
-    <Head>
-      <title>Jmember</title>
-      <meta name="viewport" content="initial-scale=1.0, width=device-width" />
-      <link rel="icon" type="image/x-icon" href="/js.png" />
-      <link rel="preconnect" href="https://fonts.googleapis.com" />
-      <link rel="preconnect" href="https://fonts.gstatic.com" />
-      <link
-        href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700;800&display=swap"
-        rel="stylesheet"
-      />
-    </Head>
-  );
-}
