+++ prettier.config.js
@@ -1,3 +1,6 @@
 module.exports = {
   plugins: [require("prettier-plugin-tailwindcss")],
+  tabWidth: 2,
+  singleQuote: false,
+  semi: true,
 };
