+++ utils/index.ts
@@ -1,0 +1,3 @@
+export const includeString = (first: string, second: string) => {
+  return first.toLowerCase().includes(second.toLowerCase());
+};
